# Introduce client-side cache in GrpcStorageProxy