--- a/optuna/storages/_grpc/api.proto
+++ b/optuna/storages/_grpc/api.proto
@@ -111,9 +111,9 @@ service StorageService {
     rpc GetTrial(GetTrialRequest) returns (GetTrialReply);
 
     /**
-     * Get all trials in a study.
+     * Get trials in a study.
      */
-    rpc GetAllTrials(GetAllTrialsRequest) returns (GetAllTrialsReply);
+    rpc GetTrials(GetTrialsRequest) returns (GetTrialsReply);
 }
 
 /**
@@ -378,17 +378,18 @@ message GetTrialReply {
 }
 
 /**
- * Request to get all trials in a study.
+ * Request to get trials in a study.
  */
-message GetAllTrialsRequest {
+message GetTrialsRequest {
     int64 study_id = 1;
-    repeated TrialState states = 2;
+    repeated int64 included_trial_ids = 2;
+    int64 trial_id_greater_than = 3;
 }
 
 /**
- * Reply to get all trials in a study.
+ * Reply to get trials in a study.
  */
-message GetAllTrialsReply {
+message GetTrialsReply {
     repeated Trial trials = 1;
 }