--- a/optuna/storages/_grpc/auto_generated/api_pb2_grpc.py
+++ b/optuna/storages/_grpc/auto_generated/api_pb2_grpc.py
@@ -126,10 +126,10 @@ class StorageServiceStub(object):
                 request_serializer=api__pb2.GetTrialRequest.SerializeToString,
                 response_deserializer=api__pb2.GetTrialReply.FromString,
                 _registered_method=True)
-        self.GetAllTrials = channel.unary_unary(
-                '/optuna.StorageService/GetAllTrials',
-                request_serializer=api__pb2.GetAllTrialsRequest.SerializeToString,
-                response_deserializer=api__pb2.GetAllTrialsReply.FromString,
+        self.GetTrials = channel.unary_unary(
+                '/optuna.StorageService/GetTrials',
+                request_serializer=api__pb2.GetTrialsRequest.SerializeToString,
+                response_deserializer=api__pb2.GetTrialsReply.FromString,
                 _registered_method=True)
 
 
@@ -282,9 +282,9 @@ class StorageServiceServicer(object):
         context.set_details('Method not implemented!')
         raise NotImplementedError('Method not implemented!')
 
-    def GetAllTrials(self, request, context):
+    def GetTrials(self, request, context):
         """*
-        Get all trials in a study.
+        Get trials in a study.
         """
         context.set_code(grpc.StatusCode.UNIMPLEMENTED)
         context.set_details('Method not implemented!')
@@ -383,10 +383,10 @@ def add_StorageServiceServicer_to_server(servicer, server):
                     request_deserializer=api__pb2.GetTrialRequest.FromString,
                     response_serializer=api__pb2.GetTrialReply.SerializeToString,
             ),
-            'GetAllTrials': grpc.unary_unary_rpc_method_handler(
-                    servicer.GetAllTrials,
-                    request_deserializer=api__pb2.GetAllTrialsRequest.FromString,
-                    response_serializer=api__pb2.GetAllTrialsReply.SerializeToString,
+            'GetTrials': grpc.unary_unary_rpc_method_handler(
+                    servicer.GetTrials,
+                    request_deserializer=api__pb2.GetTrialsRequest.FromString,
+                    response_serializer=api__pb2.GetTrialsReply.SerializeToString,
             ),
     }
     generic_handler = grpc.method_handlers_generic_handler(
@@ -888,7 +888,7 @@ class StorageService(object):
             _registered_method=True)
 
     @staticmethod
-    def GetAllTrials(request,
+    def GetTrials(request,
             target,
             options=(),
             channel_credentials=None,
@@ -901,9 +901,9 @@ class StorageService(object):
         return grpc.experimental.unary_unary(
             request,
             target,
-            '/optuna.StorageService/GetAllTrials',
-            api__pb2.GetAllTrialsRequest.SerializeToString,
-            api__pb2.GetAllTrialsReply.FromString,
+            '/optuna.StorageService/GetTrials',
+            api__pb2.GetTrialsRequest.SerializeToString,
+            api__pb2.GetTrialsReply.FromString,
             options,
             channel_credentials,
             insecure,