<!-- Thank you for creating a pull request! In general, we merge your pull request after it gets two or more approvals. To proceed to the review process by the maintainers, please make sure that the PR meets the following conditions: (1) it passes all CI checks, and (2) it is neither in the draft nor WIP state. If you wish to discuss the PR in the draft state or need any other help, please mention the Optuna development team in the PR. -->

## Motivation
<!-- Describe your motivation why you will submit this PR. This is useful for reviewers to understand the context of PR. -->

Follow-up #5852 
To make `GrpcStorageProxy` faster by introducing a client-side cache on GrpcStorageProxy.

## Description of the changes
<!-- Describe the changes in this PR. -->

This PR makes `GrpcStorageProxy` faster by cache `storage.get_all_trials` calls. Here is the performance comparison results and the benchmark script I used.

| `master` branch | This PR |
| --- | --- |
| 87.392s (+/- 0.6052104) | 48.293s (+/- 0.223478) |

```bash
# Start a MySQL server
$ docker run -d --rm -p 3306:3306 -v ./mysql-conf:/etc/mysql/conf.d -e MYSQL_USER=optuna -e MYSQL_DATABASE=optuna -e MYSQL_PASSWORD=password -e MYSQL_ALLOW_EMPTY_PASSWORD=yes --name optuna-mysql mysql:8.0

# Start a gRPC storage proxy server
$ python3 -c "import optuna; from concurrent.futures import ThreadPoolExecutor; from optuna.storages import run_grpc_proxy_server; run_grpc_proxy_server(storage=optuna.storages.get_storage('mysql+pymysql://optuna:password@127.0.0.1:3306/optuna'), host='localhost', port=13000, thread_pool=ThreadPoolExecutor(10))"
```

```python
from __future__ import annotations

from concurrent.futures import as_completed
from concurrent.futures import ProcessPoolExecutor
import threading
import time
from typing import Literal

import numpy as np

import optuna
from optuna.storages import BaseStorage
from optuna.storages import GrpcStorageProxy


study_name = "grpc-example"
n_workers = 10
n_repeat = 5
n_trials = 1000

STORAGE_KIND = Literal["grpc", "rdb"]
storage_kinds: list[STORAGE_KIND] = ["grpc"]


def get_storage(storage_kind: STORAGE_KIND) -> BaseStorage:
    if storage_kind == "grpc":
        return GrpcStorageProxy(host="localhost", port=13000)
    elif storage_kind == "rdb":
        return optuna.storages.get_storage("mysql+pymysql://optuna:password@127.0.0.1:3306/optuna")


def objective(trial: optuna.Trial) -> float:
    s = 0.0
    for i in range(10):
        s += trial.suggest_float(f"x{i}", -10, 10) ** 2
    return s


def study_optimize(storage_kind: STORAGE_KIND, n_trials: int) -> int:
    start = time.time()
    storage = get_storage(storage_kind)
    study = optuna.load_study(
        storage=storage, study_name=study_name, sampler=optuna.samplers.RandomSampler()
    )
    study.optimize(objective, n_trials=n_trials)
    elapsed = time.time() - start
    return elapsed


def bench_study_optimize(storage_kind: STORAGE_KIND) -> list[int]:
    elapsed = []
    study = optuna.create_study(storage=get_storage("rdb"), study_name=study_name)
    with ProcessPoolExecutor(n_workers) as pool:
        futures = [
            pool.submit(study_optimize, storage_kind, n_trials / n_workers)
            for _ in range(n_workers)
        ]
        for fut in as_completed(futures):
            elapsed.append(fut.result())
    return elapsed


def main():
    try:
        optuna.delete_study(storage=get_storage("rdb"), study_name=study_name)
    except Exception:
        pass

    elapsed_optimize = {}
    elapsed_get_all_trials = {}
    for storage_kind in storage_kinds:
        elapsed_optimize[storage_kind] = bench_study_optimize(storage_kind)
        optuna.delete_study(storage=get_storage("rdb"), study_name=study_name)

    for key in elapsed_optimize:
        print(f"{key=} {n_trials=} {n_repeat=} {n_workers=}")
        elapsed = elapsed_optimize[key]
        print(f"  study.optimize: {np.mean(elapsed)=} {np.std(elapsed)=}")


if __name__ == "__main__":
    main()
```