@c-bata 2024-12-20
@y0z @nabenabe0928 Could you review this PR?