@y0z 2024-12-24
The changes look good to me.

However, I sometimes encountered the following error when I ran the benchmarking script.
This problem happened for **both the master and this PR**.
Moreover, the problem occurs stochastically and is not deterministic.
This makes it difficult to reproduce.
This problem should be resolved in another PR.

```python
[I 2024-12-24 15:26:34,943] Trial 924 finished with value: 290.2282271538501 and parameters: {'x7': [I 2024-12-24 15:40:03,132] Trial 972 finished with value: 225.24752071871094 and parameters: {'x7': -6.2522912173605505, 'x2': -9.919431600233237, 'x4': 0.044403182081824255, 'x1': 4.557349591146185, 'x6': 1.7976231513999519, 'x0': 0.3857780539269626, 'x3': 6.8879029045391285, 'x9': 2.375230714328289, 'x5': -2.495022132598197, 'x8': 2.0735256279292553}. Best is trial 625 with value: 80.69101882439877.
[I 2024-12-24 15:40:03,648] Trial 973 finished with value: 289.3464461810266 and parameters: {'x7': 2.381450328633779, 'x2': 1.9933529378663124, 'x4': 2.258098349460056, 'x1': 3.975028662239829, 'x6': 1.9883090130037946, 'x0': 7.296691363420539, 'x3': -4.43790924552073, 'x9': 9.69317044115908, 'x5': -6.608645512992846, 'x8': 6.6543188068702435}. Best is trial 625 with value: 80.69101882439877.
[I 2024-12-24 15:40:04,151] Trial 974 finished with value: 206.050847349017 and parameters: {'x7': -3.851493335810887, 'x2': -3.5448847331385664, 'x4': -2.0483589374393807, 'x1': 5.714988193665674, 'x6': -8.106049930368089, 'x0': 3.675040384660633, 'x3': -3.2218795371186637, 'x9': -1.1367502366625644, 'x5': -7.13484941915987, 'x8': 0.03196690174535988}. Best is trial 625 with value: 80.69101882439877.
[I 2024-12-24 15:40:04,611] Trial 975 finished with value: 282.71786380668885 and parameters: {'x7': -5.74931015351887, 'x2': -4.28596562642862, 'x4': 4.491680141165432, 'x1': 8.65879309760668, 'x6': -8.477996393827132, 'x0': 4.373272358617948, 'x3': 1.7027574571316748, 'x9': 0.46300610792595265, 'x5': -5.839393434437474, 'x8': 2.8159722200976454}. Best is trial 625 with value: 80.69101882439877.
concurrent.futures.process._RemoteTraceback: 
"""
grpc._channel._InactiveRpcError: <_InactiveRpcError of RPC that terminated with:
        status = StatusCode.UNKNOWN
        details = "Exception calling application: An exception is raised during the commit. This typically happens due to invalid data in the commit, e.g. exceeding max length. "
        debug_error_string = "UNKNOWN:Error received from peer  {grpc_message:"Exception calling application: An exception is raised during the commit. This typically happens due to invalid data in the commit, e.g. exceeding max length. ", grpc_status:2, created_time:"2024-12-24T15:38:04.407866+09:00"}"
>

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/Users/yozaki/.asdf/installs/python/3.11.7/lib/python3.11/concurrent/futures/process.py", line 217, in _sendback_result
    result_queue.put(_ResultItem(work_id, result=result,
  File "/Users/yozaki/.asdf/installs/python/3.11.7/lib/python3.11/multiprocessing/queues.py", line 371, in put
    obj = _ForkingPickler.dumps(obj)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/yozaki/.asdf/installs/python/3.11.7/lib/python3.11/multiprocessing/reduction.py", line 51, in dumps
    cls(buf, protocol).dump(obj)
TypeError: cannot pickle '_thread.RLock' object
"""

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "/Users/yozaki/work/review/optuna/benchmark.py", line 81, in <module>
    main()
  File "/Users/yozaki/work/review/optuna/benchmark.py", line 71, in main
    elapsed_optimize[storage_kind] = bench_study_optimize(storage_kind)
                                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/yozaki/work/review/optuna/benchmark.py", line 59, in bench_study_optimize
    elapsed.append(fut.result())
                   ^^^^^^^^^^^^
  File "/Users/yozaki/.asdf/installs/python/3.11.7/lib/python3.11/concurrent/futures/_base.py", line 449, in result
    return self.__get_result()
           ^^^^^^^^^^^^^^^^^^^
  File "/Users/yozaki/.asdf/installs/python/3.11.7/lib/python3.11/concurrent/futures/_base.py", line 401, in __get_result
    raise self._exception
TypeError: cannot pickle '_thread.RLock' object
```