@nabenabe0928 2024-12-24
Hmmm, I tried with `n_trials=300`, but I could only observe degradation...

Plus, it took me so long to finish `n_trials=1000` (300+ seconds) in my environment.

I used MySQL as a database and Python 3.11 on Ubuntu 20.04.

```
$ pip freeze
# Output
alembic==1.14.0
colorlog==6.9.0
greenlet==3.1.1
grpcio==1.68.1
Mako==1.3.8
MarkupSafe==3.0.2
mysqlclient==2.2.6
numpy==2.2.1
packaging==24.2
protobuf==5.29.2
PyYAML==6.0.2
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
```