@nabenabe0928 2024-12-25
I measured the runtime of pure RDBStorage in my env and it took me about a minute for the code below.

<details>
<summary>Benchmark Code</summary>

```python
import os
import time

import optuna


def objective(trial: optuna.Trial) -> float:
    s = 0.0
    for i in range(10):
        s += trial.suggest_float(f"x{i}", -10, 10) ** 2
    
    return s


study = optuna.create_study(sampler=optuna.samplers.RandomSampler(), storage=os.environ["OPTUNA_STORAGE"])
start = time.time()
study.optimize(objective, n_trials=1000, n_jobs=10)
print(time.time() - start)

```

</detials>