@HideakiImamura 2024-12-26
I confirmed that this PR is faster than the current master by using the following benchmark code. It looks like the `get_all_trials` is completely speeded up.

- Server side code
```python
from __future__ import annotations

import os
from optuna.storages import run_grpc_proxy_server
import optuna
import signal


STORAGE_URL = "YOUR_MYSQL_URL"

def handler(signum, frame):
    print("Received signal", signum)
    exit(0)

if __name__ == "__main__":
    signal.signal(signal.SIGINT, handler)
    signal.signal(signal.SIGTERM, handler)

    run_grpc_proxy_server(
        optuna.storages.get_storage(STORAGE_URL),
        host="localhost", 
        port=13000,
    )

```

- Client side code
```python
from __future__ import annotations

import argparse
from concurrent.futures import ProcessPoolExecutor, as_completed
from collections.abc import Sequence
from collections.abc import Container
import cProfile
import time
import numpy as np
import optuna
import threading
from typing import Literal
from optuna.storages import GrpcStorageProxy, BaseStorage
import os
import uuid
from time import sleep


study_name = "grpc-example"
n_workers = 10
n_repeat = 5
n_trials = 1000

STORAGE_KIND = Literal["grpc", "rdb", "middleware"]
STORAGE_URL = "YOUR_MYSQL_URL"


def get_storage(storage_kind: STORAGE_KIND) -> BaseStorage:
    if storage_kind == "grpc":
        return GrpcStorageProxy(host="localhost", port=13000)
    elif storage_kind == "rdb":
        return optuna.storages.get_storage(STORAGE_URL)


def objective(trial: optuna.Trial) -> float:
    s = 0.0
    for i in range(10):
        trial.set_user_attr(f"attr{i}", "attr value")
    for i in range(10):
        s += trial.suggest_float(f"x{i}", -10, 10) ** 2
    return s


def study_optimize(storage_kind: STORAGE_KIND, n_trials: int) -> int:
    start = time.time()
    storage = get_storage(storage_kind)
    while True:
        try:
            study = optuna.load_study(storage=storage, study_name=study_name, sampler=optuna.samplers.RandomSampler())
            break
        except:
            sleep(10)
    study.optimize(objective, callbacks=[optuna.study.MaxTrialsCallback(n_trials)])
    elapsed = time.time() - start
    return elapsed


def bench_study_optimize(storage_kind: STORAGE_KIND) -> list[int]:
    elapsed = []
    study = optuna.create_study(
        storage=get_storage("rdb"), 
        study_name=study_name, 
        load_if_exists=True,
        sampler=optuna.samplers.RandomSampler(),
    )
    study_optimize(storage_kind, n_trials)
    return elapsed


def main(storage: STORAGE_KIND, benchmark_file: str):
    try:
        optuna.delete_study(storage=get_storage("rdb"), study_name=study_name)
    except:
        pass
    
    cProfile.runctx(
        f"bench_study_optimize('{storage}')", 
        globals(), 
        locals(), 
        filename=benchmark_file,
    )
    optuna.delete_study(storage=get_storage(storage), study_name=study_name)


if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("--storage", type=str, required=True)
    parser.add_argument("--benchmark_file", type=str, required=True)
    args = parser.parse_args()
    main(args.storage, args.benchmark_file)

```

- Results

| Current master | This PR |
| --- | --- |
| 604 sec | 203 sec |
| <img width="1005" alt="Screenshot 2024-12-26 at 10 35 09" src="https://github.com/user-attachments/assets/3f637054-26bf-431a-ace2-8751a1cbc2af" /> | <img width="999" alt="Screenshot 2024-12-26 at 10 35 29" src="https://github.com/user-attachments/assets/37305a8d-2974-4fee-baf6-673b39b31e41" /> |